% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gn.R
\name{check_gn}
\alias{check_gn}
\title{Check if GeneNetwork is live}
\usage{
check_gn(url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{url}{URL for GeneNetwork}
}
\value{
Version number (as a character string)
}
\description{
Check if GeneNetwork is live
}

