% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geno.R
\name{get_geno}
\alias{get_geno}
\title{Get genotype data}
\usage{
get_geno(cross, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{cross}{Cross name as a single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
Matrix with genotypes; rows are markers, columns are strains
}
\description{
Get genotype data for a given cross
}
\examples{
get_geno("BXD")
}

