% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gmap.R
\name{get_gmap}
\alias{get_gmap}
\title{Get genetic map}
\usage{
get_gmap(cross, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{cross}{Cross name as a single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
Data frame with marker name, chromosome, cM position, Mbp position
}
\description{
Get genetic map for a given cross
}
\examples{
get_gmap("BXD")
}
