% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_pheno}
\alias{get_pheno}
\title{Get phenotype data}
\usage{
get_pheno(dataset, trait = NULL, url = gnapi_url())
}
\arguments{
\item{dataset}{Short abbreviation for a dataset, or name of a group}

\item{trait}{Optional trait name}

\item{url}{The URL for the GeneNetwork API}
}
\value{
A data frame, with samples as rows and either traits as columns or
with the columns being detailed information for a single trait
}
\description{
Get phenotype data
}
\examples{
\donttest{ph1 <- get_pheno("B30_K_1206_R")}
ph2 <- get_pheno("BXD", "10002")
ph3 <- get_pheno("HC_M2_0606_P", "1436869_at")
}
\seealso{
\code{\link[=get_geno]{get_geno()}}
}
