% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno.R
\name{get_probeset}
\alias{get_probeset}
\title{Get probeset data}
\usage{
get_probeset(dataset, probeset,
  url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{dataset}{Name or ID for dataset as a single character string}

\item{probeset}{Name of probeset}

\item{url}{URL for GeneNetwork API}
}
\value{
Data frame with strain id, name, phenotype value, and standard error
}
\description{
Get phenotype data for a given probeset within a dataset
}
\examples{
get_probeset("HC_U_0304_R", "104617_at")
get_probeset("HC_M2_0606_P", "1443823_s_at")
}
