% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNapi-package.R
\docType{package}
\name{GNapi-package}
\alias{GNapi}
\alias{GNapi-package}
\title{GNapi: Connection to the GeneNetwork API}
\description{
Tools for connecting to the GeneNetwork API.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/GNapi/GNapi.html}{GNapi user guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqtl/GNapi}
  \item Report bugs at \url{https://github.com/rqtl/GNapi/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (0000-0002-4914-6671)

}
\keyword{internal}
