% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnapi_url.R
\name{gnapi_url}
\alias{gnapi_url}
\title{The URL for the GeneNetwork API}
\usage{
gnapi_url()
}
\value{
The URL, as character string
}
\description{
The current URL for the GeneNetwork API
}
\examples{
gnapi_url()

}
