% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_cross}
\alias{info_cross}
\title{Get info about a cross}
\usage{
info_cross(cross, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{cross}{Name of cross, as single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
A list
}
\description{
Get info about a cross
}
\examples{
info_cross("BXD")
}
