% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_datasets}
\alias{info_datasets}
\title{Metadata about all datasets for a group}
\usage{
info_datasets(group, url = gnapi_url())
}
\arguments{
\item{group}{Name of a group of datasets}

\item{url}{The URL for the GeneNetwork API}
}
\value{
A data frame
}
\description{
Get metadata about all datasets for a group
}
\examples{
g <- list_groups()
info <- info_datasets(g$Name[8])
}
