% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_geno}
\alias{info_geno}
\title{Get meta information for genotypes}
\usage{
info_geno(cross, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{cross}{Cross name as a single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
Matrix with genotypes; rows are markers, columns are strains
}
\description{
Get meta information for the genotypes for a cross
}
\examples{
info_geno("BXD")
}
