% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets for a cross}
\usage{
list_datasets(cross, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{cross}{Name of cross, as single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
A data frame with dataset ID, name, and description
}
\description{
List available datasets for a cross.
}
\examples{
list_datasets("BXD")
}
