% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{list_species}
\alias{list_species}
\title{Get list of species}
\usage{
list_species(species = NULL, url = gnapi_url())
}
\arguments{
\item{species}{Optional character string, to obtain a particular, single species}

\item{url}{URL for GeneNetwork API}
}
\value{
Data frame with columns \code{FullName}, \code{Id}, \code{Name}, and \code{TaxonomyId}.
}
\description{
Get list of available species
}
\examples{
list_species()
}
\seealso{
\code{\link[=list_groups]{list_groups()}}, \code{\link[=list_datasets]{list_datasets()}}
}
