% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stuff.R
\name{run_correlation}
\alias{run_correlation}
\title{Calculate correlations}
\usage{
run_correlation(dataset, group, trait, type = c("sample", "tissue"),
  method = c("pearson", "spearman"), n_results = 500,
  url = gnapi_url())
}
\arguments{
\item{dataset}{Name of database for the trait (\code{Short_Abbreviation})}

\item{group}{Target database name to be correlated against}

\item{trait}{ID for trait used for correlation}

\item{type}{Sample or tissue correlation}

\item{method}{Pearson or Spearman correlation}

\item{n_results}{Number of results to return}

\item{url}{The URL for the GeneNetwork API}
}
\value{
Data frame
}
\description{
Find correlated traits in GeneNetwork
}
\examples{
out <- run_correlation("HC_M2_0606_P", "BXDPublish", "1427571_at")
}
