% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stuff.R
\name{run_gemma}
\alias{run_gemma}
\title{Run gemma}
\usage{
run_gemma(dataset, trait, use_loco = FALSE, maf = 0.01,
  url = gnapi_url())
}
\arguments{
\item{dataset}{ID for dataset being used (\code{Short_Abbreviation})}

\item{trait}{ID for trait being mapped}

\item{use_loco}{Whether to use LOCO (leave one chromosome out)}

\item{maf}{Filter on minor allele frequency}

\item{url}{The URL for the GeneNetwork API}
}
\value{
A data frame
}
\description{
Perform a genome scan using gemma
}
\examples{
out <- run_gemma("BXDPublish", "10015")
}
