% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gn.R
\name{check_gn}
\alias{check_gn}
\title{Check if GeneNetwork is live}
\usage{
check_gn(url = gnapi_url())
}
\arguments{
\item{url}{URL for GeneNetwork}
}
\value{
Character string.
}
\description{
Check if GeneNetwork is live
}
\examples{
check_gn()
}
