% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_geno}
\alias{get_geno}
\title{Get genotype data}
\usage{
get_geno(group, url = gnapi_url(), format = "geno")
}
\arguments{
\item{group}{Name of group}

\item{url}{The URL for the GeneNetwork API}

\item{format}{The group's genotypes format}
}
\value{
A data frame
}
\description{
Get genotype data
}
\examples{
g <- get_geno("QSM")
}
\seealso{
\code{\link[=get_pheno]{get_pheno()}}, \code{\link[=list_groups]{list_groups()}}
}
