% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno.R
\name{get_pheno}
\alias{get_pheno}
\title{Get phenotype data}
\usage{
get_pheno(dataset, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{dataset}{Dataset name or ID as a single character string.}

\item{url}{URL for GeneNetwork API}
}
\value{
Data frame with strain ID, strain name, phenotype value, and standard error.
}
\description{
Get phenotype data for a dataset
}
\examples{
get_pheno("10001")
}
