% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_dataset}
\alias{info_dataset}
\title{Get info about a dataset}
\usage{
info_dataset(dataset, url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{dataset}{Name of dataset, as single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
A list
}
\description{
Get info about a dataset
}
\examples{
info_dataset("HC_U_0304_R")
}

