% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_dataset}
\alias{info_dataset}
\title{Metadata about a dataset}
\usage{
info_dataset(dataset, url = gnapi_url())
}
\arguments{
\item{dataset}{The short abbreviation for the dataset}

\item{url}{The URL for the GeneNetwork API}
}
\value{
A data frame with a single row
}
\description{
Get metadata about a dataset
}
\examples{
g <- list_groups()
d <- list_datasets(g$Name[8])
info <- info_dataset(d$Short_Abbreviation[1])
}
