% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_pheno}
\alias{info_pheno}
\title{Get summary information about a phenotype}
\usage{
info_pheno(group, trait, url = gnapi_url())
}
\arguments{
\item{group}{Name of a group of datasets}

\item{trait}{Trait identifier (can be a vector)}

\item{url}{The URL for the GeneNetwork API}
}
\value{
A data frame
}
\description{
Get summary information about a phenotype
}
\examples{
info_pheno("BXD", "10002")
info_pheno("HC_M2_0606_P", "1436869_at")
}
