% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets for a group}
\usage{
list_datasets(group, url = gnapi_url())
}
\arguments{
\item{group}{Name of group, as single character string}

\item{url}{URL for GeneNetwork API}
}
\value{
A data frame with dataset ID, name, and description
}
\description{
List available datasets for a group.
}
\examples{
list_datasets("BXD")
}
\seealso{
\code{\link[=list_species]{list_species()}}, \code{\link[=list_groups]{list_groups()}}
}
