% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{list_groups}
\alias{list_groups}
\title{Get list of groups}
\usage{
list_groups(url = gnapi_url())
}
\arguments{
\item{url}{URL for GeneNetwork API}
}
\value{
Data frame with columns \code{FullName}, \code{Id}, \code{Name}, and \code{TaxonomyId}.
}
\description{
Get list of available groups
}
\examples{
list_groups()
}
\seealso{
\code{\link[=list_species]{list_species()}}, \code{\link[=list_datasets]{list_datasets()}}
}
