% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{list_species}
\alias{list_species}
\title{Get list of species}
\usage{
list_species(url = "http://test-gn2.genenetwork.org/api_pre1/")
}
\arguments{
\item{url}{URL for GeneNetwork API}
}
\value{
Data frame with number, species common name and formal name.
}
\description{
Get list of available species
}
\examples{
list_species()
}

