% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stuff.R
\name{run_rqtl}
\alias{run_rqtl}
\title{Run R/qtl}
\usage{
run_rqtl(dataset, trait, method = c("hk", "ehk", "em", "imp", "mr",
  "mr-imp[", "mr-argmax"), model = c("normal", "binary", "2part", "np"),
  n_perm = 0, control_marker = NULL, interval_mapping = FALSE,
  url = gnapi_url())
}
\arguments{
\item{dataset}{ID for dataset being used (\code{Short_Abbreviation})}

\item{trait}{ID for trait being mapped}

\item{method}{Indicates whether to use the EM algorithm, imputation, Haley-Knott regression, the extended Haley-Knott method, or marker regression.}

\item{model}{The phenotype model: the usual normal model, a model for binary traits, a two-part model, or non-parametric analysis}

\item{n_perm}{Number of permutations}

\item{control_marker}{Name of marker to use as control}

\item{interval_mapping}{Whether to use interval mapping}

\item{url}{The URL for the GeneNetwork API}
}
\value{
A data frame
}
\description{
Perform a genome scan using R/qtl
}
\examples{
out <- run_rqtl("BXDPublish", "10015")
}
